requirements.txt
Flask==3.1.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-CORS==6.0.0
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.0.1
reportlab==4.0.4  # Para gerar PDFs
stripe==6.0.0  # Para pagamentos
python-dateutil==2.8.2
pdfkit==1.0.0  # Para converter HTML para PDF
Jinja2==3.1.6  # Para renderizar templates HTML
